reorganize component structure
