--- app/_actions.ts
@@ -1,17 +1,0 @@
-'use server';
-import { revalidatePath } from 'next/cache';
-
-const clearCachesByServerAction = async (path?: string) => {
-  try {
-    if (path) {
-      revalidatePath(path);
-    } else {
-      revalidatePath('/');
-      revalidatePath('/[lang]');
-    }
-  } catch (error) {
-    console.error('clearCachesByServerAction=> ', error);
-  }
-};
-
-export default clearCachesByServerAction;
