+++ providers/Providers.tsx
@@ -4,7 +4,7 @@ import { useState, type ReactElement } from 'react';
 import { httpBatchLink, loggerLink } from '@trpc/client';
 import { trpc } from '~/app/_trpc/client';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { SessionProvider } from '~/contexts/SessionPrivider';
+import { SessionProvider } from '~/providers/SessionPrivider';
 import type { Session } from 'lucia';
 
 export default function Providers({
